Move view model to separate file
